task watcher
Summary:
This is the first diff for later.  Its a task watcher.
It provides basic task watching and restarting behavior.

Tasks can be watched at any time the Watcher is still active.
If the watcher is canceled, all tasks inside it are canceled also.

Canceled tasks are awaited, by default there is a 5m timeout on Watcher waiting
for cancels, this is configurable in the Watcher init.

Reviewed By: ambv

Differential Revision: D9651948

fbshipit-source-id: 1c81b86b2c8642ad67c466f522f21de74b80a6ea