later.Watcher() pre-exit callbacks
Summary:
Pre-exit callbacks, allow a user of Watcher() to trigger some callback
right before the Watcher `__aexit__` exits.  This allows for behavior
to occure before the watcher starts canceling tasks.

This is useful in cases where the Watcher context reguardless of if its going to
raise and Exception or not and we need to perform some action before Tasks are
canceled and waited on, which could take time.

Reviewed By: sdunster

Differential Revision: D10500274

fbshipit-source-id: 89fefa2d112fa90287d917058ba3d1daea2cc563