task watcher unittests to run under later.unittest.TestCase
Summary:
Moved Watcher to be tested under the later TestCase, and discovered a place where we were not cleaning up tasks we created in watcher.  So since I take the position that no task should ever be "forgotten" without checking its value, I fixed those things.  And converted the rests of the tests.

I had to fix one of the tests because it created a task that could not be cancelled, and the underlying IsolatedAsyncioTestCase, waits for all tasks after canceling them.  So I changed it up to just be slow after being canceled, but if canceled twice should await fine.

Reviewed By: thatch

Differential Revision: D18308403

fbshipit-source-id: 6f6cb35c361724b39eb08d84fcd1b14aea341e18