TaskTask use saved coro_repr for __repr__ and ignore some Tasks
Summary:
Tasks that are done() and have a result of None, should be considered managed.  thrift-py3 does this in servers when it calls rpc methods, we don't have a way for the thrift server to monitor the python task since its results are stored in a C++ Folly Promise object.

Also the coro object can be garbage collected before the end of the test which makes understanding left over task errors very hard. So use the coro_repr we save in the `__init__` to make `__repr__` more useful.

Reviewed By: cooperlees

Differential Revision: D24550663

fbshipit-source-id: a8a4e20f1fbc670b75d4f1c4dc9c4f289c3a9123