use stdlib instead of backports after 3.8
Summary:
Later was built on backports from 3.8 to give us features even while using 3.7.
When we moved to 3.8 we kept them because I wasn't exactly sure at the time that all the fixes we baked into 3.8 async_mock had made it into the stdlib 3.8 yet.

Well if the version of python is greater than 3.8 lets just use the stdlib

Bumping version so we can release to pypi

Differential Revision: D43885595

fbshipit-source-id: d5e1ca89984b818c512cea48e9dee0ebc9040b3a