allow watcher to be used as a task_group
Summary: There has been some talk about getting a TaskGroup before py3.11.  Watcher is pretty much a TaskGroup but finished tasks was not anticipated all tasks were expected to go on forever.    So added a done_ok flag for watcher that defaults to "True" so we can see if anything actually needs that behavior.

Reviewed By: dcsommer

Differential Revision: D44723981

fbshipit-source-id: 64a0721c63d5fca79e5183376abed886905abaa3