Revert D51681518: Multisect successfully blamed "D51681518: [later] cinder compatability for differences in all_tasks" for otest failure
Summary:
This diff is reverting D51681518
Depends on D51696082
D51681518: [later] cinder compatability for differences in all_tasks by fried has been identified to be causing the following test failure:

Tests affected:
- [fblearner/flow/recurring_service/tests:test_server - fblearner.flow.recurring_service.tests.test_server.TestRecurringService: test_error_handling](https://www.internalfb.com/intern/test/281475062751785/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3637357
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: fried

Differential Revision: D51696084

fbshipit-source-id: c99379a6037af9b533a06ccfc4038f0eab05f395