disable a test since during coverage runs it blows up
Summary:
I'm not sure how to fix this one.  The issue is with an error that occurs in asyncio outside of a test. But the test was expected to fail.  We do want to look for asnycio issues that happen after a test otherwise we would never know about weird ways people break the eventloop.

For now lets just disable it.

Also add some nocovers around python version gated code

Reviewed By: itamaro

Differential Revision: D58595923

fbshipit-source-id: 43e54a02b89a4f135d9f85f680392406204b624a