fix TestCase chopping testmethod in traceback
Summary:
I don't understand it completely but this has something to do with https://fburl.com/code/69ink2vd and its use of __unittest in globals.  As soon as I moved the TestTask to a new file the results of errors in tests are as expected.

Before this in 3.12 at least the traceback stopped at `super().result()` in `TestTask.result`

Reviewed By: itamaro

Differential Revision: D74109988

fbshipit-source-id: 3220e4a514df1ac03ac616078d3a52b939437790