gather as a coroutine so its safe to use with asyncio.run
Summary:
Having gather be a coroutine simplifies many old use cases where people use asyncio.gather and immediately pass it to `loop.run_until_complete`.

With this any asyncio.gather that only deals with coroutines can be converted so it can be directly passed to asyncio.run

The typing BS alone makes this worthy to save in later.

Also cleaned up type checking for 3.12
Moved test_coroutine.py out of the unittest directory since it was placed in the wrong directory.

Bumped Version

Reviewed By: itamaro

Differential Revision: D78189067

fbshipit-source-id: c0868131144036198285782e8ad29410ba815819